08-GenAI-Tracing/requirements.txt
python-dotenv
openai
azure-identity
azure-ai-projects
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-openai
azure-monitor-opentelemetry-exporter
opentelemetry-exporter-otlp